fix frequency int overflow

+++ src/hallelujah.h
@@ -34,7 +34,7 @@ struct HallelujahWord {
         : translation_(translation), ipa_(ipa), frequency_(frequency) {}
     std::vector<std::string> translation_;
     std::string ipa_;
-    int frequency_;
+    double frequency_;
 };
 
 class HallelujahEngine;
